libstore: split up derivations files into `derivation/` dir
Following up on splitting out `derivation/full-inputs.hh`, split a
few more things out of `derivations.{cc,hh}`:

- `derivation/output.{cc,hh}`: the `DerivationOutput` type, its
  `path` methods, and associated maps.

  Simply because it's big enough.

- `derivation/aterm.{cc,hh}`: the ATerm parser and unparser, the
  derivation-hash-modulo computation (`DrvHashModulo` and friends,
  which is defined in terms of the ATerm unparsing), and the wire
  protocol read/write functions.

  Because we want to separate conceptual data from specific
  serialization formats.

- `derivation/json.cc`: the JSON serialization.

  Again, because we want to separate conceptual data from specific
  serialization formats.

`derivations.hh` includes `derivation/output.hh` (which it needs for
`DerivationT` itself), but deliberately does *not* re-export
`derivation/aterm.hh`; the files that actually use the ATerm /
hash-modulo / wire-protocol functions include it directly.

Two previously file-local helpers had to be newly exposed, because the
split cuts between them and their callers in `derivations.cc`:

- `hasDynamicDrvDep`, now in `derivations.hh`: a general predicate on
  a derivation's inputs map (not ATerm-specific), used by
  `checkInvariants`.

- `hashDerivationModuloImpl<maskOutputs>`, now declared in
  `derivation/aterm.hh` with explicit instantiations in `aterm.cc`:
  `processDerivationOutputPaths` needs the masked (`true`) variant.

  It's certainly not good to newly expose something with "Impl" in the
  name, but this will be fixed in a future refactor. It is better to do
  the split first.
