+++ src/libstore-tests/derivation-advanced-attrs.cc
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include "nix/store/derivations.hh"
+#include "nix/store/derivation/aterm.hh"
 #include "nix/store/derived-path.hh"
 #include "nix/store/derivation-options.hh"
 #include "nix/store/globals.hh"
