+++ src/libstore-tests/derivation-parser-bench.cc
@@ -1,5 +1,6 @@
 #include <benchmark/benchmark.h>
 #include "nix/store/derivations.hh"
+#include "nix/store/derivation/aterm.hh"
 #include "nix/store/store-api.hh"
 #include "nix/util/tests/test-data.hh"
 #include "nix/store/store-open.hh"
