+++ src/libstore/include/nix/store/derivations.hh
@@ -7,12 +7,14 @@
 #include "nix/store/content-address.hh"
 #include "nix/util/repair-flag.hh"
 #include "nix/store/derivation/full-inputs.hh"
+#include "nix/store/derivation/output.hh"
 #include "nix/store/derived-path-map.hh"
 #include "nix/store/parsed-derivations.hh"
 #include "nix/util/sync.hh"
 #include "nix/util/variant-wrapper.hh"
 
 #include <boost/unordered/concurrent_flat_map_fwd.hpp>
+#include <ranges>
 #include <variant>
 
 namespace nix {
@@ -28,132 +30,6 @@ class Store;
 /* Abstract syntax of derivations. */
 
 namespace derivation {
-
-/**
- * A single output of a `Basic` (and `Full`) derivation.
- */
-struct Output
-{
-    /**
-     * The traditional non-fixed-output derivation type.
-     */
-    struct InputAddressed
-    {
-        StorePath path;
-
-        bool operator==(const InputAddressed &) const = default;
-        auto operator<=>(const InputAddressed &) const = default;
-    };
-
-    /**
-     * Fixed-output derivations, whose output paths are content
-     * addressed according to that fixed output.
-     */
-    struct CAFixed
-    {
-        /**
-         * Method and hash used for expected hash computation.
-         *
-         * References are not allowed by fiat.
-         */
-        ContentAddress ca;
-
-        /**
-         * Return the \ref StorePath "store path" corresponding to this output
-         *
-         * @param drvName The name of the derivation this is an output of, without the `.drv`.
-         * @param outputName The name of this output.
-         */
-        StorePath path(const StoreDirConfig & store, std::string_view drvName, OutputNameView outputName) const;
-
-        bool operator==(const CAFixed &) const = default;
-        auto operator<=>(const CAFixed &) const = default;
-    };
-
-    /**
-     * Floating-output derivations, whose output paths are content
-     * addressed, but not fixed, and so are dynamically calculated from
-     * whatever the output ends up being.
-     * */
-    struct CAFloating
-    {
-        /**
-         * How the file system objects will be serialized for hashing
-         */
-        ContentAddressMethod method;
-
-        /**
-         * How the serialization will be hashed
-         */
-        HashAlgorithm hashAlgo;
-
-        bool operator==(const CAFloating &) const = default;
-        auto operator<=>(const CAFloating &) const = default;
-    };
-
-    /**
-     * Input-addressed output which depends on a (CA) derivation whose hash
-     * isn't known yet.
-     */
-    struct Deferred
-    {
-        bool operator==(const Deferred &) const = default;
-        auto operator<=>(const Deferred &) const = default;
-    };
-
-    /**
-     * Impure output which is moved to a content-addressed location (like
-     * CAFloating) but isn't registered as a realization.
-     */
-    struct Impure
-    {
-        /**
-         * How the file system objects will be serialized for hashing
-         */
-        ContentAddressMethod method;
-
-        /**
-         * How the serialization will be hashed
-         */
-        HashAlgorithm hashAlgo;
-
-        bool operator==(const Impure &) const = default;
-        auto operator<=>(const Impure &) const = default;
-    };
-
-    typedef std::variant<InputAddressed, CAFixed, CAFloating, Deferred, Impure> Raw;
-
-    Raw raw;
-
-    bool operator==(const Output &) const = default;
-    auto operator<=>(const Output &) const = default;
-
-    MAKE_WRAPPER_CONSTRUCTOR(Output);
-
-    /**
-     * Force choosing a variant
-     */
-    Output() = delete;
-
-    /**
-     * \note when you use this function you should make sure that you're
-     * passing the right derivation name. When in doubt, you should use
-     * the safer interface provided by
-     * BasicDerivation::outputsAndOptPaths
-     */
-    std::optional<StorePath>
-    path(const StoreDirConfig & store, std::string_view drvName, OutputNameView outputName) const;
-};
-
-typedef std::map<std::string, Output> Outputs;
-
-/**
- * These are analogues to the previous `Outputs` data type, but they
- * also contains, for each output, the (optional) store path in which
- * it would be written. To calculate values of these types, see the
- * corresponding functions in `Basic`.
- */
-typedef std::map<std::string, std::pair<Output, std::optional<StorePath>>> OutputsAndOptPaths;
 
 /**
  * For inputs that are sub-derivations, we specify exactly which
@@ -380,11 +256,6 @@ template<>
 void Full::checkInvariants(Store & store) const;
 
 /**
- * Print a derivation.
- */
-std::string unparse(const Full & drv, const StoreDirConfig & store);
-
-/**
  * Determine whether this derivation should be resolved before building.
  *
  * Resolution is needed when:
@@ -475,7 +346,6 @@ Full parseJsonAndValidate(Store & store, const nlohmann::json & json);
 
 using BasicDerivation = derivation::Basic;
 using Derivation = derivation::Full;
-using DerivationOutput = derivation::Output;
 
 /**
  * Compute the store path that would be used for a derivation without writing it.
@@ -485,13 +355,19 @@ using DerivationOutput = derivation::Output;
 StorePath computeStorePath(const StoreDirConfig & store, const Derivation & drv);
 
 /**
- * Read a derivation from a file.
+ * Does the derivation have a dependency on the output of a dynamic
+ * derivation?
+ *
+ * In other words, does it on the output of derivation that is itself an
+ * output of a derivation? This corresponds to a dependency that is an
+ * inductive derived path with more than one layer of
+ * `DerivedPath::Built`.
  */
-Derivation parseDerivation(
-    const StoreDirConfig & store,
-    std::string && s,
-    std::string_view name,
-    const ExperimentalFeatureSettings & xpSettings = experimentalFeatureSettings);
+template<std::ranges::input_range Range>
+bool hasDynamicDrvDep(Range && drvsMap)
+{
+    return std::ranges::any_of(std::forward<Range>(drvsMap), [](auto & kv) { return !kv.second.childMap.empty(); });
+}
 
 /**
  * \todo Remove.
@@ -510,107 +386,6 @@ bool isDerivation(std::string_view fileName);
 std::string outputPathName(std::string_view drvName, OutputNameView outputName);
 
 /**
- * The hashes modulo of a derivation.
- *
- * Each output is given a hash, although in practice only the content-addressed
- * derivations (fixed-output or not) will have a different hash for each
- * output.
- */
-struct DrvHashModulo
-{
-    /**
-     * Single hash for the derivation
-     *
-     * This is for an input-addressed derivation that doesn't
-     * transitively depend on any floating-CA derivations.
-     */
-    using DrvHash = Hash;
-
-    /**
-     * Known CA drv's output hashes, for fixed-output derivations whose
-     * output hashes are always known since they are fixed up-front.
-     */
-    using CaOutputHashes = std::map<std::string, Hash>;
-
-    /**
-     * This derivation doesn't yet have known output hashes.
-     *
-     * Either because itself is floating CA, or it (transtively) depends
-     * on a floating CA derivation.
-     */
-    using DeferredDrv = std::monostate;
-
-    using Raw = std::variant<DrvHash, CaOutputHashes, DeferredDrv>;
-
-    Raw raw;
-
-    bool operator==(const DrvHashModulo &) const = default;
-    // auto operator <=> (const DrvHashModulo &) const = default;
-
-    MAKE_WRAPPER_CONSTRUCTOR(DrvHashModulo);
-};
-
-/**
- * Returns hashes with the details of fixed-output subderivations
- * expunged.
- *
- * A fixed-output derivation is a derivation whose outputs have a
- * specified content hash and hash algorithm. (Currently they must have
- * exactly one output (`out`), which is specified using the `outputHash`
- * and `outputHashAlgo` attributes, but the algorithm doesn't assume
- * this.) We don't want changes to such derivations to propagate upwards
- * through the dependency graph, changing output paths everywhere.
- *
- * For instance, if we change the url in a call to the `fetchurl`
- * function, we do not want to rebuild everything depending on it---after
- * all, (the hash of) the file being downloaded is unchanged.  So the
- * *output paths* should not change. On the other hand, the *derivation
- * paths* should change to reflect the new dependency graph.
- *
- * For fixed-output derivations, this returns a map from the name of
- * each output to its hash, unique up to the output's contents.
- *
- * For regular derivations, it returns a single hash of the derivation
- * ATerm, after subderivations have been likewise expunged from that
- * derivation.
- */
-DrvHashModulo hashDerivationModulo(Store & store, const Derivation & drv);
-
-/**
- * If a derivation is input addressed and doesn't yet have its input
- * addressed (is deferred) try using `hashDerivationModulo`.
- *
- * Does nothing if not deferred input-addressed, or
- * `hashDerivationModulo` indicates it is missing inputs' output paths
- * and is not yet ready (and must stay deferred).
- */
-void resolveInputAddressed(Store & store, Derivation & drv);
-
-struct DrvHashFct
-{
-    using is_avalanching = std::true_type;
-
-    std::size_t operator()(const StorePath & path) const noexcept
-    {
-        return std::hash<std::string_view>{}(path.to_string());
-    }
-};
-
-/**
- * Memoisation of hashDerivationModulo().
- */
-typedef boost::concurrent_flat_map<StorePath, DrvHashModulo, DrvHashFct> DrvHashes;
-
-// FIXME: global, though at least thread-safe.
-extern DrvHashes drvHashes;
-
-struct Source;
-struct Sink;
-
-Source & readDerivation(Source & in, const StoreDirConfig & store, BasicDerivation & drv, std::string_view name);
-void writeDerivation(Sink & out, const StoreDirConfig & store, const BasicDerivation & drv);
-
-/**
  * This creates an opaque and almost certainly unique string
  * deterministically from the output name.
  *
@@ -627,8 +402,6 @@ std::string hashPlaceholder(const OutputNameView outputName);
 constexpr unsigned expectedJsonVersionDerivation = 4;
 
 } // namespace nix
-
-JSON_IMPL_WITH_XP_FEATURES(nix::DerivationOutput)
 
 namespace nlohmann {
 template<typename Inputs>
