+++ src/libstore/include/nix/store/meson.build
@@ -34,7 +34,9 @@ headers = [ config_pub_h ] + files(
   'content-address.hh',
   'daemon.hh',
   'derivation-options.hh',
+  'derivation/aterm.hh',
   'derivation/full-inputs.hh',
+  'derivation/output.hh',
   'derivations.hh',
   'derived-path-map.hh',
   'derived-path.hh',
