+++ src/libstore/meson.build
@@ -279,6 +279,9 @@ sources = files(
   'content-address.cc',
   'daemon.cc',
   'derivation-options.cc',
+  'derivation/aterm.cc',
+  'derivation/json.cc',
+  'derivation/output.cc',
   'derivations.cc',
   'derived-path-map.cc',
   'derived-path.cc',
